basic vm testing
